requirements.txt
flask==3.1.0
python-telegram-bot==21.10
google-genai==1.14.0
python-dotenv==1.1.0
gunicorn==23.0.0
